Move rlp::encode(vector) into encode_vector.hpp
