+++ core/silkworm/rlp/encode_vector.hpp
@@ -14,32 +14,36 @@
    limitations under the License.
 */
 
-#include "receipt.hpp"
+#ifndef SILKWORM_RLP_ENCODE_VECTOR_HPP_
+#define SILKWORM_RLP_ENCODE_VECTOR_HPP_
 
-#include <silkworm/common/util.hpp>
-#include <silkworm/rlp/encode_vector.hpp>
+#include <vector>
+
+#include <silkworm/rlp/encode.hpp>
 
 namespace silkworm::rlp {
 
-static Header header(const Receipt& r) {
-    Header h;
-    h.list = true;
-    h.payload_length = 1;
-    h.payload_length += length(r.cumulative_gas_used);
-    h.payload_length += length(full_view(r.bloom));
-    h.payload_length += length(r.logs);
-    return h;
+template <class T>
+size_t length(const std::vector<T>& v) {
+    size_t payload_length{0};
+    for (const T& x : v) {
+        payload_length += length(x);
+    }
+    return length_of_length(payload_length) + payload_length;
 }
 
-void encode(Bytes& to, const Receipt& r) {
-    if (r.type != Transaction::Type::kLegacy) {
-        to.push_back(static_cast<uint8_t>(r.type));
+template <class T>
+void encode(Bytes& to, const std::vector<T>& v) {
+    Header h{true, 0};
+    for (const T& x : v) {
+        h.payload_length += length(x);
     }
-    encode_header(to, header(r));
-    encode(to, r.success);
-    encode(to, r.cumulative_gas_used);
-    encode(to, full_view(r.bloom));
-    encode(to, r.logs);
+    encode_header(to, h);
+    for (const T& x : v) {
+        encode(to, x);
+    }
 }
 
 }  // namespace silkworm::rlp
+
+#endif  // SILKWORM_RLP_ENCODE_VECTOR_HPP_
