+++ core/silkworm/state/state.hpp
@@ -22,6 +22,7 @@
 #include <silkworm/common/base.hpp>
 #include <silkworm/types/account.hpp>
 #include <silkworm/types/block.hpp>
+#include <silkworm/types/receipt.hpp>
 
 namespace silkworm {
